updated quality index calculation to avoid int type mismatch

+++ polarspike/binarizer.py
@@ -5,7 +5,6 @@ This module makes heavy use of the Polars library to increase performance.
 @ Marvin Seifert 2024
 """
 
-
 import polars as pl
 from functools import partial
 import numpy as np
@@ -13,7 +12,7 @@ import scipy.signal as signal
 
 
 def timestamps_to_binary_polars(
-    df_timestamps: pl.DataFrame, sample_rate: int = None, max_window: int = None
+        df_timestamps: pl.DataFrame, sample_rate: int = None, max_window: int = None
 ) -> pl.DataFrame:
     """
     This function converts timestamps to binary signals. It takes a DataFrame with timestamps and converts them to binary
@@ -85,7 +84,7 @@ def timestamps_to_binary_polars(
 
 
 def fill_missing_repeats(
-    max_repeat: int, nr_bins: int, df_result: pl.DataFrame
+        max_repeat: int, nr_bins: int, df_result: pl.DataFrame
 ) -> pl.DataFrame:
     """
     This function fills missing repeats with zeros. This is necessary to have a consistent DataFrame that is all
@@ -126,7 +125,7 @@ def fill_missing_repeats(
                 schema={
                     "cell_index": pl.Int64,
                     "value": pl.UInt8,
-                    "repeat": pl.Int32,
+                    "repeat": pl.Int64,
                     "index": pl.Int64,
                 },
             )
@@ -138,7 +137,7 @@ def fill_missing_repeats(
 
 
 def apply_on_group(
-    sample_rate: int, max_window: int, group_df: pl.DataFrame
+        sample_rate: int, max_window: int, group_df: pl.DataFrame
 ) -> pl.DataFrame:
     """
     Apply the timestamps_to_binary_polars function on a group of a DataFrame.
@@ -166,7 +165,7 @@ def apply_on_group(
 
 
 def timestamps_to_binary_multi(
-    df: pl.DataFrame, bin_size: int, max_window: int, max_repeat: int
+        df: pl.DataFrame, bin_size: int, max_window: int, max_repeat: int
 ) -> pl.DataFrame:
     """
     Convert timestamps to binary signals for multiple cells.
@@ -303,7 +302,7 @@ def calc_tradqi(kernel_fits: np.ndarray) -> float:
 
 
 def kernel_template(
-    width: float = 0.0100, sampling_freq: float = 17852.767845719834
+        width: float = 0.0100, sampling_freq: float = 17852.767845719834
 ) -> np.ndarray:
     """
     Create Gaussian kernel by providing the width (FWHM) value. According to wiki, this is approx. 2.355*std of dist.
@@ -328,7 +327,7 @@ def kernel_template(
     gtime = np.arange(-k, k)
 
     # create Gaussian window
-    gauswin = np.exp(-(4 * np.log(2) * gtime**2) / fwhm**2)
+    gauswin = np.exp(-(4 * np.log(2) * gtime ** 2) / fwhm ** 2)
     gauswin = gauswin / np.sum(gauswin)
 
     # initialize filtered signal vector
